# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ITFY E-Voting Platform API
  description: |
    # ITFY E-Voting Platform Backend API
    
    Welcome to the comprehensive API documentation for the ITFY E-Voting platform.
    This platform enables event organizers to create and manage voting events with 
    candidates, categories, and secure payment-based voting.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **Event Management**: Create, manage, and publish voting events
    - **Category Management**: Organize candidates into voting categories
    - **Candidate Management**: Register and manage candidates with profiles
    - **Voting System**: Purchase vote bundles and cast votes securely
    - **Payment Integration**: Paystack payment gateway integration
    - **Forms & Submissions**: Dynamic nomination forms with approval workflows
    - **Analytics**: Comprehensive voting analytics and reporting
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## User Roles
    - **super_admin**: Full system access
    - **admin**: Administrative access to events and users
    - **organiser**: Can create and manage events
    - **moderator**: Can view and moderate content
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
...
    
    ## Pagination
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    CandidateAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated candidates
...
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
...
    SearchParam:
      in: query
      name: search
      schema:
        type: string
      description: Search query string
...
paths:
...
  /events/public:
    get:
      tags:
        - Events
      summary: List public events
      description: Retrieve published public events
      operationId: listPublicEvents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /events/upcoming:
    get:
      tags:
        - Events
      summary: Get upcoming events
      description: Retrieve upcoming events
      operationId: getUpcomingEvents
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Upcoming events retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
  /events/slug/{slug}:
    get:
      tags:
        - Events
      summary: Get event by slug
      description: Retrieve event using URL-friendly slug
      operationId: getEventBySlug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
          description: Event slug
      responses:
        '200':
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events:
    post:
      tags:
        - Events
      summary: Create a new event
      description: Create a new event (Admin/Organiser only)
      operationId: createEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    get:
      tags:
        - Events
      summary: List all events (Admin)
      description: Retrieve paginated list of all events
      operationId: listEvents
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [active, upcoming, archived, cancelled, pending]
        - in: query
          name: visibility
          schema:
            type: string
            enum: [public, private, unlisted]
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /events/{id}:
    get:
      tags:
        - Events
      summary: Get event by ID
      description: Retrieve detailed information about a specific event
      operationId: getEventById
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Events
      summary: Update event
      description: Update an event (Admin/Organiser only)
      operationId: updateEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Events
      summary: Delete event
      description: Soft delete an event (Admin only)
      operationId: deleteEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/publish:
    put:
      tags:
        - Events
      summary: Publish event
      description: Publish an event to make it visible
      operationId: publishEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/unpublish:
    put:
      tags:
        - Events
      summary: Unpublish event
      description: Unpublish an event to hide it
      operationId: unpublishEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event unpublished successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/status:
    put:
      tags:
        - Events
      summary: Update event status
      description: Change event status
      operationId: updateEventStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [active, upcoming, archived, cancelled, pending]
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/cancel:
    put:
      tags:
        - Events
      summary: Cancel event
      description: Cancel an event
      operationId: cancelEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
      responses:
        '200':
          description: Event cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/complete:
    put:
      tags:
        - Events
      summary: Complete event
      description: Mark an event as completed
      operationId: completeEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Event completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/featured:
    put:
      tags:
        - Events
      summary: Toggle featured status
      description: Toggle event featured status (Admin only)
      operationId: toggleEventFeatured
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Featured status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/categories:
    get:
      tags:
        - Events
      summary: Get event categories
      description: Retrieve categories for an event
      operationId: getEventCategories
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Categories retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/candidates:
    get:
      tags:
        - Events
      summary: Get event candidates
      description: Retrieve all candidates for an event
      operationId: getEventCandidates
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Candidates retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/stats:
    get:
      tags:
        - Events
      summary: Get event statistics
      description: Retrieve statistics for an event (Admin only)
      operationId: getEventStats
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalCategories:
                        type: integer
                      totalCandidates:
                        type: integer
                      totalVotes:
                        type: integer
                      totalRevenue:
                        type: number
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/votes/summary:
    get:
      tags:
        - Events
      summary: Get event vote summary
      description: Retrieve vote summary for an event (Admin only)
      operationId: getEventVoteSummary
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Vote summary retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/results:
    get:
      tags:
        - Events
      summary: Get event results
      description: Retrieve voting results for an event
      operationId: getEventResults
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '403':
          description: Results not yet published
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/results/publish:
    put:
      tags:
        - Events
      summary: Publish event results
      description: Publish voting results for an event (Admin only)
      operationId: publishEventResults
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Results published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{id}/duplicate:
    post:
      tags:
        - Events
      summary: Duplicate event
      description: Create a copy of an event (Admin/Organiser only)
      operationId: duplicateEvent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name for the duplicated event
      responses:
        '201':
          description: Event duplicated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==================
  # Category Endpoints
  # ==================
...
  /votes/results/event/{eventId}:
    get:
      tags:
        - Votes
      summary: Get event results
      description: Retrieve voting results for an event
      operationId: getVoteEventResults
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
...
  /votes/count/event/{eventId}:
    get:
      tags:
        - Votes
      summary: Get event vote count
      description: Retrieve total votes in an event
      operationId: getVoteCountByEvent
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Count retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
...
  /votes/event/{eventId}:
    get:
      tags:
        - Votes
      summary: Get votes by event
      description: Retrieve votes for an event (Admin only)
      operationId: getVotesByEvent
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Votes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vote'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /bundles/event/{eventId}:
    get:
      tags:
        - Bundles
      summary: Get bundles by event
      description: Retrieve bundles for an event
      operationId: getBundlesByEvent
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bundles retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bundle'
...
  /payments/event/{eventId}:
    get:
      tags:
        - Payments
      summary: Get payments by event
      description: Retrieve payments for an event (Admin only)
      operationId: getPaymentsByEvent
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Payments retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /forms/event/{eventId}:
    get:
      tags:
        - Forms
      summary: Get forms by event
      description: Get all forms for an event (Admin only)
      operationId: getFormsByEvent
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: Event ID
        - name: form_type
          in: query
          schema:
            type: string
            enum: [nomination, registration, survey, feedback]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, closed, archived]
      responses:
        '200':
          description: List of forms for event
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Form'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /submissions/event/{eventId}:
    get:
      tags:
        - Submissions
      summary: Get submissions by event
      description: Get all submissions for an event (Admin only)
      operationId: getSubmissionsByEvent
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, duplicate, under_review]
      responses:
        '200':
          description: Submissions for event
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Submission'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /slides/event/{eventId}:
    get:
      tags:
        - Slides
      summary: Get slides by event
      description: Get slides for an event (Public)
      operationId: getSlidesByEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Slides for event
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Slide'
...